+++ CHANGELOG.md
@@ -10,6 +10,7 @@
             - proper ProgressBar?
             - update battery, temperatures etc
             - migrate BatchCollection
+    - fixed "Reset FPGA" button
 - 2024-08-29 4.1.1
     - ensure metadata 'stomped' by plugins appear in exports
     - fix interpolation defaults and persistence